third_party/WebKit/LayoutTests/dom/attr/set-attribute-node-from-iframe-expected.txt

ownerDocument of Attr should be set on setAttributeNode

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS attr.ownerDocument is iframeDocument
PASS attr.ownerDocument is document
PASS attr.ownerDocument is iframeDocument
PASS successfullyParsed is true

TEST COMPLETE
